Move the implementation of Wayland specific methods into a separate file
